+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.1.8
+* 
+
 ### v1.1.7
 * Actor : Added tracer argument to fire_bullet
 * Damager : Added TRACER enum
@@ -6,7 +9,7 @@
     * Allows for calling gm functions with wrapped arguments, which will automatically unwrap.
     * Additionally, can call with `:` syntax using wrapped Instance objects (e.g., `actor:skill_util_unlock_cooldown(skill)`, with `actor` being a wrapped Actor object).
         * Basically there is no longer a need to do `.value`.
-    * Removed dedicated actor:recalculate_stats
+    * Removed dedicated `actor:recalculate_stats`.
 
 ### v1.1.6
 * Fixed Survivor add_skin and add_skill